feat(sysuser): refactor user update logic and improve tests
- Move update logic from controller to service layer
- Change return type of update API to R<Void>
- Fix role assignment logic in user update
- Update test cases to use SysUserDto and JsonUtil
- Adjust orgId field type from Integer to Long
- Add updateUser method in SysUserService interface and implementation

+++ src/main/java/org/huazhi/system/sysuser/entity/SysUser.java
@@ -61,7 +61,7 @@ public class SysUser extends PanacheEntityBase {
 	/**
 	 * 组织ID
 	 */
-	private Integer orgId;
+	private Long orgId;
 
 	/**
 	 * 锁定时间
